feat: Local implicit modules for @guppy (#105)
Makes `@guppy` define local modules by default. Fixes #101.

- Removes `_Guppy.set_module`
- Adds a `_Guppy.compile` that compiles and returns the local module.
- Similarly, a `_Guppy.take_module` that returns the module without
compiling it (and removes it from the local context).
- Adds a `@guppy(compile=True)` option as a fall-back to the previous
behaviour, where it directly compiles a Hugr.

Look at the first commit. The second is just noise updating the tests to
use the `compile` flag.

I wonder if it'd be better to have a different decorator that compiles
directly to hugrs instead of the `compile` flag.