chore: Add release-please workflow (#138)
Add release action. This should check commits to main for the string
"release-as: X.X.X" and open a PR for the release, creating a changelog
from our conventional commits.

When this is merged, it will add a git tag and github release pointing
to the merged release PR.

I've made it also check the branch `release-testing` so that we can
check if it works! After testing, I'll open a PR to get rid of this line