feat: Upgrade Hugr and start using the shared Pydantic model (#201)
This PR makes Guppy depend on Hugr at main as a path for upgrading to
v0.4. Closes #198

Most of the edits are related to changes in the Pydantic model. However,
we also had to get rid of `TypeApply` since this no longer exists in
Hugr. Therefore, Guppy now rejects all programs that use polymorphic
functions as values if the type arguments cannot be inferred