feat: Compile places and struct field accesses to Hugr (#291)
See #295 for context, the tests are at #293

The idea is to implement a similar logic to one used for linearity
checking in #290. As soon as we assign a variable with struct type, we
immediately unpack the struct to make the individual fields available.
We only repack once the whole struct value is requested.