chore: Turn result function into an AST node (#327)
Fixes #320.

We were relying on a hack where the `CustomCallChecker` for the
`result(tag, value)` function inserted an extra type arg for the tag
that wasn't declared in the signature of `result`. This now leads to
problems in #321

For now, the only way to fix this is to add a new node to the AST that
handles compilation of results with tags.

In the future, we could solve this more elegantly by for example adding
`Literal` types which would allow us to specify the tag type arg in
argument position:

```python
def result[n, T](tag: Literal[n], value: T) -> None: ...
```