feat: Allow py expressions in type arguments (#515)
Closes #513

Function declarations and struct definitions now also store the
`python_scope` in which they were defined in oder to resolve
py-expressions ocurring in their signatures.

Also I needed to update the way the python scope of functions is
computed: The approach using `__closure__` doesn't capture variables
that are bound in the signature. Instead we now also inspect the calling
frame to extract the `f_globals` and `f_locals` from there. The same is
also used for structs since Python classes don't have `__closure__`