feat: Cache accessed source files (#551)
We used to store the entire source code that was used to build an AST as
a field within *every node* of the AST. This was a hack so that we can
render the source line from an AST node when an error occurs. With our
new diagnostics system, we should do better than that!

With this PR, we cache the source of every file accessed by the compiler
in a new `SourceMap` class. Then, we can look up the source associated
with a span when rendering a diagnostic.

Note that we can't rely on Pythons `linecache` since we also need to
keep track of cells in jupyter notebooks as individual "files".