feat: conventional results post processing (#593)
Post processing of result streams, for converting to traditional
distributions over bitstrings if some tagging convention is used,
including conversion to a pytket BackendResult.


## Convention
Tags are assumed to be a name of a bit register unless they fit the
regex pattern `^([a-z][\w_]*)\[(\d+)\]$` (like `reg[12]`) in which case
they are assumed to refer to the nth element of a bit register.

For results of the form 
```
result("<register>", value)
```
`value` can be `{0, 1}`, wherein the register is assumed to be length 1,
or lists over those values, wherein the list is taken to be the value of
the entire register.

For results of the form
```
result("<register>[n]", value)
```
`value` can only be `{0, 1}`.
The register is assumed to be at least `n+1` in size and unset elements
are assumed to be `0`.

Subsequent writes to the same register/element in the same shot will
overwrite.

To convert to a `BackendResult` all registers must be present in all
shots, and register sizes cannot change between shots.

## TODO 
Clear documentation of the above conventions and examples.