feat: Generic function definitions (#618)
Allow function definitions that are generic over parameters of kind
`type` or `nat`.

* Store currently available type parameters in the type checking context
* Refactor type parsing logic to explicitly opt in to add free variables
to the parameter mapping
* Add a new `GenericParamValue` AST node to represent usages of generic
nat params `n` inside function bodies. We need a way to encode this in
Hugr. For now, we just emit a dummy node. See
https://github.com/CQCL/hugr/issues/1629

Note: Nested generic functions are not supported yet, so we don't have
to worry about scoping of type params.

Closes #522