feat: result appending post-processing (#713)
To enable e.g. all the elements of an array to be collated in to the
same bitstrings, see tests

Flattens lists so can be used once entire arrays can be printed too.

So some guppy that looks like for a given shot:

```python
vals = (1, 0, 1, 0)
for v in vals:
    result("c", v)
```


results in post-processed output, over 100 shots
```python
Counter({(("c", 1010"),): 100})
```