feat: Make `True` and `False` branches unconditional (#740)
For example, when building a CFG for `if True: ...`, build an
unconditional branch. This improves linearity checking for functions
involving `while True` loops, since dataflow analysis now understands
that such loops can only exit on an explicit `break` or `return`.

It also means that Guppy is now better at detecting unreachable code.
~For now, we error out in this case~. In the future it would be nice to
emit a warning when we detect this (see #739).

One complication introduced by this is that Guppy can now generate CFGs
where the exit block is unreachable. This lead to some problems with
borrowed variables. This edge case now needs to be specially handled.

I recommend reviewing each of the three commits separately