fix: Prevent reordering of operations with side-effects (#855)
Closes #853
 
This is a temporary solution until we properly solve #778. Therefore,
the implementation is somewhat hacky as we're going to remove this in
the future anyways.

The approach is to monkey-patch the `Hugr.add_node` method to implicitly
insert state-order edges between nodes that could have side-effects. The
list of side-effectful nodes is hard-coded in the compiler.

Also relies on the fix from https://github.com/CQCL/hugr/pull/1971, so
we should wait to merge until `hugr-py` is released