fix: Fix `result` / `panic` /  `exit` calls in comptime functions (#949)
Closes #919

There are two issues with these functions:

1) The checker not knowing it can treat comptime arguments as static
values - this fix is sort of a halfway between a general solution
(variables passed to Guppy function calls inside comptime functions are
now `ComptimeVariables` which just inherits from `Variable` with an
additional field to pass argument values directly) and hack (it is only
used for this particular case so isn't thought out completely)

2) As they are built-ins defined with a custom checker and no type
annotations (as the inputs could have different types), `func.ty` is
empty so you can't iterate through the function input types to update
inouts - since as far as I can see any function where this is the case
also won't need to update inouts, the current fix is to just not do that
part of the input types are empty, but this seems a bit dodgy. Maybe
there could be a way of allowing or forcing custom checker to provide a
function signature, but not sure if this is more of a follow up and this
fix is fine since it does make these functions work for now.