feat!: Remove explicit Guppy modules (#983)
Closes #956 and closes #957. See #936 for more context.

Most of the diff is coming from changing all the tests to no longer use
modules. The significant changes are the following:

* Instead of tracking definitions by module, we now have a global
`DEF_STORE` where every `@guppy` decorated definition is registered.
* For each defininition, this store also remembers the stack frame in
which it was defined. When encountering a name while checking a
function, we look it up in this stack frame and see if it corresponds to
another Guppy definition. The old "name -> definition" mapping is gone,
it's all based on what's in scope in the defining frame. See
`guppylang.checker.core.Globals` for the new lookup mechanism.
* This also means that parsing, checking, and compilation of definitions
is now on demand: We begin by checking the user-proveded entry point
(via a new `guppy.compile` method) and only check other definitions if
they are used by the entry point.
* This is driven via a new `CompilationEngine` class (defined in
`guppylang/engine.py`). It maintains a worklist of which definitions
still need to be checked/compiled and replaces the old
`GuppyModule.check` and `GuppyModule.compile` logic.

BREAKING CHANGE: Explicit `GuppyModule` declarations are no longer
possible. Instead, use the regular `@guppy` decorator everywhere without
passing an explicit module. Compilation is now triggered via the
`guppy.compile` function, passing the to-be-compiled function as an
argument.