fix: Fix invalid Hugr when reassigning a used variable (#1026)
Fixes #1023

After linearity checking, we refine BB signatures to refer to places
instead of variables. However, when looking up places for the BB inputs,
we searched in the scope of the current BB instead of it's parent where
the input should be coming from. So, if we write

```python
use(a)   # a coming from predecessor BB has type int
a = 1.0  # now float
# So at the end of this BB's scope, a has type int
```

we mistakenly assigned `float` to the `a` input instead of `int`. The
fix is easy: Just use `scope.parent_scope` when computing the BB input
signature.