docs: add quantum and qsystem gate definitions (#912)
Adding some gate definitions and doctests to the qsystem and quantum
docs

added docstrings for everything along with matrix definitions of the
gates (as in pytket
[OpType](https://docs.quantinuum.com/tket/api-docs/optype.html))

Theres quite a lot of freedom in convention for how we respresent this
(captialisation etc). Would be interested in your thoughts.

Regarding functional gates I've just mentioned that they are the same as
the std.quantum variants with different syntax. Would crosslink to the
docs but would rather link to the offfical docs when they are public.

closes #857

---------

Co-authored-by: Alec Edgington <54802828+cqc-alec@users.noreply.github.com>