chore!: Require a context when converting types to Hugr (#1022)
This is required as a precusor to #1018 to handle monomorphisation: When
compiling to Hugr, some bound type variables will need to be replaced
with a monotype. We will need access to the compiler context to look
this information up.

Instead of requiring the whole compiler context, I added a new
`ToHugrContext` protocol in which we can put everything that is needed
during the Hugr translation of types (empty for now but will be filled
in a subsequent PR). The main motivation for this is that the context
needs to change when we go under a binder.

BREAKING CHANGE: All `to_hugr` methods on types, arguments, and
parameters now require a `ToHugrContext`
BREAKING CHAGNE: `CompileableDef.compile_outer` now requires a
`ToHugrContext`
BREAKING CHANGE: `guppy.hugr_op` now passes the compiler context to the
function generating the op