feat: Add generic `guppy.const_var` declarations (#1034)
Adds a new `guppy.const_var` declaration that generalises
`guppy.nat_var` to arbitrary copyable+droppable types. This is part of
#1033.

For basic testing, I also updated type argument parsing to accept `bool`
and `float` const literals. However, the current solution for this is
not ideal (see #1030), but should be addressed in a separate PR.

Also note that Hugr lowering of these const generics is not implemented
in this PR, so the integration tests currently only check the code
instead of compiling to Hugr. Hugr compilation via monomorphisation will
follow in a separate PR.