feat!: Lower const generics via monomorphization (#1035)
Lowers the const generic introduceds in #1034 to Hugr via
monomorphization. Part of #1033.

* In `definition.common`, we add new `MonomorphizableDef` abstract base
classes to mark definitions that need to be lowered via
monomorphisation. `CheckedFunctionDef` is changed to implement
`MonomorphizableDef` instead of `CompileableDef`. All other definitions
are unchanged.

* The `checker.core.CompilerContext` is the main driver for
monomorphisation:
* When calling `CompilerContext.build_compiled_def`, we now also need to
pass the type arguments for the definition we want to compile.
* If the to be compiled function is a `MonomorphizableDef`, the context
decides which of its parameters will need to be monomorphised. This is
done by the `requires_monomorphization` function. Currently, those are
exactly the non-nat const parameters of the function (note that nat
params can be lowered to Hugr bounded nats, so monomorphization is not
needed)
* From the provided type arguments, the context picks the subset that
correspond to these to-be-monomorphized parameters, which is done via
the `partially_monomorphize_args` function. We represent these partial
monomorphizations using the `PartiallyMonomorphizedArgs` type.
* A `mono_args: PartiallyMonomorphizedArgs` is a sequence of `Argument |
None`, so that `mono_args[i] == None` means that the argument with index
`i` is not part of the monomorphization. Otherwise, `mono_args[i]`
specifies the monomorphic instantiation for parameter `i`. So a function
with 3 parameters where no monomorphization is required at all, will
have `mono_args = (None, None, None)`.
* The worklist and compilation store in the `CompilerContext` are now
indexed via a pair of `DefId` and optional `PartiallyMonomorphizedArgs`.
This way, the worklist can contain the same definition multiple times,
but with different monomorphized args.
* The same also applies to the `global_funcs` stored in the context. By
taking potential mono-arguments into account, we can create multiple
global functions corresponding to the same `GlobalConstId`. This is used
in `CustomFunctionDef.load_with_args` to support custom functions with
params that require monomorphization.

* The context also has two new methods: `type_var_to_hugr` and
`const_var_to_hugr`. Those are used when compiling types and type
arguments to Hugr. The context stores the current monomorphisation of
the function that is being compiled, and when we request to lower a
variable to Hugr, the context makes sure that the de Bruijn indices are
down-shifted to account for the monomorphized parameters.
* Similarly, `ExprCompiler.visit_GenericParamValue` looks up the value
corresponding to the current monomorphization in the context

BREAKING CHANGE: `CheckedFunctionDef` now implements
`MonomorphizableDef` instead of `CompileableDef`
BREAKING CHANGE: `CompilerContext.build_compiled_def` now requires an
instantiation for the definition's type parameters
BREAKING CHANGE: The `ToHugrContext` protocol now requires two
additional methods: `type_var_to_hugr` and `const_var_to_hugr`
BREAKING CHANGE: `CompilerContext.{compiled, worklist}` and
`CompilationEngine.compiled` are now indexed by a tuple of `DefId` and
optional `PartiallyMonomorphizedArgs`

---------

Co-authored-by: Alan Lawrence <alan.lawrence@cambridgequantum.com>