fix(emulator): pass runtime option to selene (#1174)
Closes #1173

The test doesn't really check the bug - would succeed without the fix,
it is more of a drive-by

The actual behaviour is somewhat subtle to test since you have to use
selene logging to check the runtime behaviour. If the reviewer feels
strongly this should be included I will investigate.

Unit tests for the emulator module are in progress.