fix!: Fix `nat` constructor in comptime functions (#1258)
Fixes #1250

The problem was that the imported `nat` is a class declared using
`@extend_type`. This decorator didn't wrap the result into a
`GuppyDefinition`, so was not equipped with comptime behaviour.

I changed the `@extend_type` decorator to now return `GuppyObject`s by
default. Note that we still need an escape hatch to return the unwrapped
class to support the `array` type whose comptime behaviour is customised
to act like lists. This can be achieved by setting the new
`return_class=True` flag in the decotator.

Note that this change is not breaking for `guppylang` since the
deprecated `@guppy.extend_type` decorator still behaves as before.

drive-by: Added a `run_nat_function` pytest fixture to test nat
functions.

BREAKING CHANGE: `guppylang_internals.decorator.extend_type` now returns
a `GuppyDefinition` by default. To get the previous behaviour of
returning the annotated class unchanged, pass `return_class=True`.