Weaken selene constraint
This is strictly wrong: we are now pulling in an out of date version
(wrt. tket versions, serialisation of wasm extension) of selene for
guppy, meaning that execution tests will fail and the
guppylang.emulator submodule will be broken.

However, this is necessary for selene to pull in guppy as a testing
dependency, without pulling in a specific version which clashes with
the version we're trying to test.