fix: compilation of affine-bounded type variables (#1308)
fixes #1306

The first part of the fix is to make affine-bounded type parameters
(declared `[T : Drop]`) compile to Hugr TypeParams bounded by linear
rather than Copyable; this fixes the error in the issue (invalid
instantiation of Copyable-bounded function with type-arg
`borrow_array<....>`), but raises a different problem: guppy does not
insert the necessary `drop` operations, thus leading to
`UnconnectedPort` validation errors (as the ports are now linear).

This was because the `requires_drop` function used when inserting drops
(#1090) inspects the `hugr_bound` property of type variables which also
needs updating that affine-in-guppy things should become linear-in-Hugr.
In this area, update some docs (RIP equatable), refactor using
`super()`, and standardize use of `ht` for `hugr.tys`.

Note that (contrary to python-semver-checks) this is not breaking: the
APIs removed (optional extra parameter to TypeParam, and settable
TypeParam::is_affine field) were only added in #1166, which has not yet
been released.