docs: Rework the repeat-until-success example notebook (#1370)
Reworks the repeat-until-success scheme example notebook to clarify what
is going on in the code and extends the experiments done with the Guppy
function.

Includes an image similar to another notebook (swap_test.ipynb),
compatible with `guppy-docs` rendering.

> [!IMPORTANT]  
> When `guppy-docs` is updated to include this PR, the path must be
updated in the page index as the file name changed. This is prepared in
https://github.com/Quantinuum/guppy-docs/pull/32.

Closes #1368